refactor(web): cleanup after typed-view-models migration
- Correct SmartyTemplateRule path in View.php docblock.
- Drop the no-op looksLikeJsInsideScript stub; parseTemplate already
  constrains matches to Smarty bodies so the escape hatch was dead.
- Harden YourAccountView min_pass_len against a null Config value by
  casting to int in the handler.

+++ web/pages/page.youraccount.php
@@ -46,7 +46,7 @@ $youraccountView = new \Sbpp\View\YourAccountView(
     user_aid: (int) $userbank->GetAid(),
     web_permissions: BitToString($userbank->GetProperty("extraflags")),
     server_permissions: SmFlagsToSb($userbank->GetProperty("srv_flags")),
-    min_pass_len: MIN_PASS_LENGTH,
+    min_pass_len: (int) MIN_PASS_LENGTH,
 );
 
 $theme->setLeftDelimiter('-{');
